* Split background items and tests into own files
