+++ src/core/background/furniture.js
@@ -1,0 +1,24 @@
+/**
+ * @copyright BBC 2019
+ * @author BBC Children's D+E
+ * @license Apache-2.0
+ */
+import fp from "../../../lib/lodash/fp/fp.js";
+import { isSprite, addSprite } from "./sprite.js";
+import { isImage, addImage } from "./image.js";
+import { isSpine, addSpine } from "./spine.js";
+import { isText, addText } from "./text.js";
+import { isParticles, addParticles } from "./particles.js";
+
+export const furnish = scene => () => {
+    const configs = scene.context.theme.furniture || [];
+    const conditionPairs = [
+        [isSpine(scene), addSpine(scene)],
+        [isImage(scene), addImage(scene)],
+        [isSprite(scene), addSprite(scene)],
+        [isParticles(scene), addParticles(scene)],
+        [isText, addText(scene)],
+    ];
+    const dispatcher = fp.cond(conditionPairs);
+    configs.forEach(dispatcher);
+};
